tf_assert

A thin wrapper around tf$Assert() that automatically constructs an informative error message (passed on to data argument), which includes the expression passed to condition, the values of the symbols found in the expression, as well as the full R call stack at the time the tf$Assert() node is created.

tf_assert(condition, ..., expr = substitute(condition),
  summarize = NULL, name = NULL)

Arguments

condition
A boolean tensor

...
Additional elements passed on to data. (e.g, an informative error message as a string, additional tensor values that might be useful to have in the error message, etc.)

expr
A language object, provided in case condition is already computed prior to the call

summarize
Print this many entries of each tensor.

name
A name for this operation (optional).

Examples

x <- tf$constant(-1)
try(tf_assert(x > 0, "oopsies! x must be greater than 0"))

#> Error in py_call_impl(callable, dots$args, dots$keywords) : 
#>   InvalidArgumentError: Expected 'tf.Tensor(False, shape=(), dtype=bool)' to be true. Summarized data: <Assert condition R expression>: x > 0
#> oopsies! x must be greater than 0
#> `x` value:
#> -1.0
#> R call stack:
#> <R call 1>: tf_assert(x > 0, "oopsies! x must be greater than 0")
#> <R call 2>: doTryCatch(return(expr), name, parentenv, handler)
#> <R call 3>: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> <R call 4>: tryCatchList(expr, classes, parentenv, handlers)
#> <R call 5>: tryCatch(expr, error = function(e) {
#> <R call 6>: try(tf_assert(x > 0, "oopsies! x must be greater than 0"))
#> <R call 7>: eval(expr, envir, enclos)
#> <R call 8>: eval(expr, envir, enclos)
#> <R call 9>: withVisible(eval(expr, envir, enclos))
#> <R call 10>: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler, error = eHandler, message = mHandler)
#> <R call 11>: doTryCatch(return(expr), name, parentenv, handler)
#> <R call 12>: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> <R call 13>: tryCatchList(expr, classes, parentenv, handlers)
#> <R call 14>: tryCatch(expr, error = function(e) {
#> <R call 15>: try(f, silent = TRUE)
#> <R call 16>: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler, error = eHandler, message = mHandler))
#> <R call 17>: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler, error = eHandler, message = mHandler)))
#> <R call 18>: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, debug = debug, last = i == length(out), use_try = stop_on_error != 2L, keep_warning = keep_warning, keep_message = keep_message, output_handler = output_handler, include_timing = include_timing)
#> <R call 19>: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> <R call 20>: highlight_examples(code, topic, env = env)
#> <R call 21>: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()), topic = tools::file_path_sans_ext(topic$file_in), run_examples = examples, run_dont_run = run_dont_run)
#> <R call 22>: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> <R call 23>: .f(.x[[i]], ...)
#> <R call 24>: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run)
#> <R call 25>: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run, seed = seed, override = override, preview = FALSE, devel = devel)
#> <R call 26>: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run, seed = seed, lazy = lazy, override = override, preview = preview, devel = devel)
#> <R call 27>: pkgdown::build_site(...)
#> <R call 28>: (function (..., crayon_enabled, crayon_colors, pkgdown_internet) 
#> <R call 29>: (function (what, args, quote = FALSE, envir = parent.frame()) 
#> <R call 30>: do.call(do.call, c(readRDS("/var/folders/zc/63kn9zds1qs4pk6v8h4q8mwc0000gn/T//Rtmp2Xh11M/file14f5577d48fff"), list(envir = .GlobalEnv)), envir = .GlobalEnv)
#> <R call 31>: saveRDS(do.call(do.call, c(readRDS("/var/folders/zc/63kn9zds1qs4pk6v8h4q8mwc0000gn/T//Rtmp2Xh11M/file14f5577d48fff"), list(envir = .GlobalEnv)), envir = .GlobalEnv), file = "/var/folders/zc/63kn9zds1qs4pk6v8h4q8mwc0000gn/T//Rtmp2Xh11M/file14f552439ebcf")
#> <R call 32>: withCallingHandlers({
#> <R call 33>: doTryCatch(return(expr), name, parentenv, handler)
#> <R call 34>: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> <R call 35>: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> <R call 36>: doTryCatch(return(expr), name, parentenv, handler)
#> <R call 37>: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]), names[nh], parentenv, handlers[[nh]])
#> <R call 38>: tryCatchList(expr, classes, parentenv, handlers)
#> <R call 39>: tryCatch(withCallingHandlers({
#> 
#> Detailed traceback: 
#>   File "/usr/local/opt/python@2/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/tensorflow_core/python/util/tf_should_use.py", line 198, in wrapped
#>     return _add_should_use_warning(fn(*args, **kwargs))
#>   File "/usr/local/opt/python@2/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/tensorflow_core/python/ops/control_flow_ops.py", line 156, in Assert
#>     (condition, "\n".join(data_str)))
#>