tf.case

This is a minimal wrapper around tf.case() that allows you to supply the pred_fn_pairs using the ~.

tf_case(..., pred_fn_pairs = list(...), default = NULL,
  exclusive = FALSE, name = "case")

Arguments

..., pred_fn_pairs
a list pred_fn_pairs supplied with the ~ like so: pred ~ fn_body

default
a function, optionally specified with the ~, (or something coercible to a function via as.function())

exclusive
bool, whether to evaluate all preds and ensure only one is true. If FALSE (the default), then the preds are evaluated in the order supplied until the first TRUE value is encountered (effectively, acting as an if()... else if() ... else if() ... chain)

name
a string, passed on to tf.case()

Value

The result from tf$case()

Examples

fizz_buzz_one <- function(x) {
  tf_case(
    x %% 15 == 0 ~ "FizzBuzz",
    x %%  5 == 0 ~ "Buzz",
    x %%  3 == 0 ~ "Fizz",
    default = ~ tf$as_string(x, precision = 0L)
  )
}

fn <- tf_function(autograph(function(n) {
  for(e in tf$range(n))
    tf$print(fizz_buzz_one(e))
}))

x <- tf$constant(16)
fn(x)

#> Error in py_call_impl(callable, dots$args, dots$keywords): RuntimeError: Evaluation error: value for ‘i’ not found.
#> 
#> Detailed traceback: 
#>   File "/usr/local/opt/python@2/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/tensorflow_core/python/eager/def_function.py", line 480, in __call__
#>     result = self._call(*args, **kwds)
#>   File "/usr/local/opt/python@2/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/tensorflow_core/python/eager/def_function.py", line 526, in _call
#>     self._initialize(args, kwds, add_initializers_to=initializers)
#>   File "/usr/local/opt/python@2/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/tensorflow_core/python/eager/def_function.py", line 431, in _initialize
#>     *args, **kwds))
#>   File "/usr/local/opt/python@2/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/tensorflow_core/python/eager/function.py", line 2258, in _get_concrete_function_internal_garbage_collected
#>     graph_function, _, _ = self._maybe_define_function(args, kwargs)
#>   File "/usr/local/opt/python@2/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/tensorflow_core/python/eager/function.py", line 2566, in _maybe_define_function
#>     graph_function = self._create_graph_function(args, kwargs)
#>   File "/usr/local/opt/python@2/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/tensorflow_core/python/eager/function.py", line 2455, in _create_graph_function
#>     capture_by_value=self._capture_by_value),
#>   File "/usr/local/opt/python@2/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/tensorflow_core/python/framework/func_graph.py", line 942, in func_graph_from_py_func
#>     func_outputs = python_func(*func_args, **func_kwargs)
#>   File "/usr/local/opt/python@2/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/tensorflow_core/python/eager/def_function.py", line 374, in wrapped_fn
#>     return weak_wrapped_fn().__wrapped__(*args, **kwds)
#>   File "<string>", line 4, in fn
#>   File "/Library/Frameworks/R.framework/Versions/3.6/Resources/library/reticulate/python/rpytools/call.py", line 21, in python_function
#>     raise RuntimeError(res[kErrorKey])